separateProject: improvments using not UV processing if not necessary and unsafe code

+++ Assets/Votyra/Core/ImageSamplers/InterpolatedUVPostProcessorStep.cs
@@ -4,7 +4,7 @@ namespace Votyra.Core.ImageSamplers
 {
     public class InterpolatedUVPostProcessorStep : ITerrainUVPostProcessor
     {
-        private readonly float _subdivision;
+        private readonly int _subdivision;
 
         public InterpolatedUVPostProcessorStep(IInterpolationConfig interpolationConfig)
         {
